Bot Reg. test teamplate
Use this POSTMAN template to test the authentication related issues or Azure Bot services app registration issues with AAD. Just replace the appID and Password from the Body section of the template after importing to POSTMAN, and make a request. You will find the issue related to Auth/ app registration in the response.